breathing: don't allow to run more than one thread breathing: stop servo if aren't breathing denexapp_config: set correct values in 'Speech' robot: connect breathing module add speech_markup, 60% are ready speech_test: use speech_markup to test speech module minor bugfixes

+++ myservo_test.py
@@ -3,4 +3,4 @@ pin = input('Enter pin')
 a = my_servo.Servo(int(pin))
 while True:
     val = input('Enter value')
-    a.start(int(val))
+    a.start(float(val))
